v5.17.5
Most patch updates are from
https://github.com/AdityaGarg8/linux-t2-patches, although the WiFi ones
are ones I found on asahi's tree.

This moves back to patching the upstream kernel, so I may have messed
something up, make sure you have a way to go back to an older kernel if
you try this before other people do.

Additionally, there's `update.sh`, which may be of use to others. It
should update the PKGBUILD and check that patches apply right.