bluetooth wip from sven at asahilinux
This has not been tested at all on intel macs, don't expect it to work.

I don't know what the firmware needs to be called, but the driver should
tell you the filename when it is modprobed. I presume it gets put in the
same folder as the wifi firmware.